refactor(schemas): split the jsonb-type file
split the jsonb-type file
