+++ packages/schemas/src/foundations/index.ts
@@ -1,2 +1,2 @@
 export * from './schemas.js';
-export * from './jsonb-types.js';
+export * from './jsonb-types/index.js';
