+++ packages/schemas/src/foundations/jsonb-types/phrases.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export type Translation = {
+  [key: string]: string | Translation;
+};
+
+export const translationGuard: z.ZodType<Translation> = z.lazy(() =>
+  z.record(z.string().or(translationGuard))
+);
